+++ backend/downloads/serializers.py
@@ -1,0 +1,21 @@
+from rest_framework import serializers
+
+from .models import Download
+
+
+class DownloadSerializer(serializers.ModelSerializer):
+    """
+    Serializer for Download model.
+    """
+
+    class Meta:
+        model = Download
+        fields = [
+            "id",
+            "title",
+            "status",
+            "file",
+            "file_size",
+            "created",
+            "updated",
+        ]