+++ changelog.md
@@ -5,6 +5,7 @@
 - 11:10 - Frontend: added "Delete" button in bookmark edit modal, wired up with backend.
 - 13:00 - Backend: created `Download` model. Updated download task.
 - 14:30 - Backend: fixed issue when downloaded video file wasn't deleted with the model instance.
+- 17:15 - Backend: Added Download model serializer.
 
 ## 21.04.2023, Fri